Fixed real_robot orientation update from sensor
This makes the real_robot free-flyer update in ``MCGlobalController::run``
consistent with ``Robot::posW``.

Note that the quaternion of the free-flyer gives the rotation from the
base-link frame to the world frame. This is *opposite* to the Plücker transform
convention.

See https://gite.lirmm.fr/multi-contact/mc_vrep/merge_requests/10#note_6863