Make a copy of the element before handling request
This fixes an issue when the element stores data to pass to the callback
and removes itself in the callback.

An example of this exists in QPSolver.cpp

This looks something like:
void foo(const std::vector<int> & v)
{
  /* snip */
  removeCategory({"A"});
  std::vector<int> v2 {...};
  addElement({"A"}, Button("B", [v2]() { foo(v2); }));
}

In such a case the vector stored in the lambda is deleted after the
removeCategory call and things will go wrong.