Correctly pass X_b_s when creating a contact
This fixes a long standing bug where the frame of the contact was not
defined as the frame of the surface but rather the frame of the parent
body. This does not matter much in the common case but this leads to
surprising behaviour when one wants to free some DoF of the contact.