[log] Fix a GUI issue
Things would not work as expected when "key" and "key_suffix" are in
the log and "key" exists as data (e.g. "key" is a double). This patch
fixes that. It also change the selector behaviour for that key so that
selecting "key" in that scenario does not select all children.