[CoPTask] Fixed broken tracking "disabling"
The main reason why this "security" was problematic is that it didn't call
AdmittanceTask::update(). Therefore, `refVelB_` would stay at its previous
value, which did not make sense and could actually be dangerous.

Now AdmittanceTask::update() is always called, and the target wrench always
updated. If the body detaches (pressure < 0), the target torque in the air will
be zero. This is consistent with measurements on HRP4LIRMM (measured torques
for the swing foot are roughly zero).