[samples][LIPMStabilizer] Minor tweaks for JVRC
- Pause for 2 seconds before starting the stabilizer to allow choreonoid
simulation to stabilize with the feet on the ground
- Aim the walking trajectory 5mm above ground to minimize early-impacts
that are not handled by this sample. Make steps of 10cm both forward and
backward. 20cm steps almost puts JVRC leg in a singular configuration.
- Slow down the put down motion for the feet

Note that the JVRC choreonoid project has also been updated separately
to start with the robot in halfsitting configuration