[BodySensor] Do not create an invalid sensor by default
Changes the behaviour when no default body sensor is provided by the
robot module. Formerly it would create an empty (invalid) BodySensor and
fill its linear/angular/... data and log it as if it was a valid sensor.
This was fine in the past when we weren't doing much with the body
sensors, but now we have state estimators, automatic logging of all body
sensors, plugins, etc. Thus we should no longer create empty body
sensors, but properly handle the case where there is no sensor, notably:

- Log all body sensors in a consistent way (do not handle the default
body sensor separately)
- Requires a valid default body sensor for the KinematicInertial
observer
- Do not fill the corresponding ROS messages if there is no bodysensor
- Ignore setting bodysensor data from MCGlobalController if it does not
exist