[utils] mc_rtc_new_fsm_controller specifies robots by name
Closes #219

- Handle robots by name instead of indices following !214 MR
- Fixes a minor bug where FSM contact loading did not throw if contacts
were specified but invalid, and silently replaced the whole contact set
with an empty set. E.g the following entry (missing `r1`) would not report an error,
and an empty contact list would end up being used instead:

```
contacts:
- r2: ground
  r1Surface: LeftFoot
  r2Surface: AllGround
```

The expected behaviour here is to throw an error "r1 entry does not
exist", as the user explicitely intended to create this contact, but
made a syntax mistake.