Report the path to the Configuration object that generated an exception
This is (almost) free when the exception is caught and silenced but
relatively expensive otherwise as we need to look for the faulty value
in the whole document.

Generally this is fine as uncaught exception will lead to the program
termination and unsilenced exception as barely used in the framework and
normally with small objects.