[mc_rbdyn][test] Reset the calibrator before ZMP test
The ZMP computation unit test assumes that the measured force lies on a
surface under an ideal force sensor that is defined at the model
position, and whose measurements are perfect.

However, when force sensor calibration files are present for the JVRC1
robot on the system where the test is being run (generated by the
new mc_force_sensor_calibration controller), the unit test would fail,
due to the influence of the calibration parameters.

This commit provides a way to reset the calibration parameters to their
default state where they have no effect on the wrenchWithoutGravity such
that it is consistent with the expectations of this unit test (this
could also prove useful in the calibration controller).

The differences are quite minor between the computed ZMP and expected ZMP (of the order of millimiters).

Here is the results of the test using the following calibration
parameters for the left force sensor:

calibdata.leftFootForceSensor: 13 parameters in that order: mass (1), rpy for X_f_ds (3), position for X_p_vb (3), wrench offset (6)

```
1.48995 # mass
0.0244621
-0.0184112
-0.0121862
0.0295351
0.00252571
-0.0685183
0.0417241
0.0198496
0.00379788
0.205764
0.379153
-0.205693
```

```
Test command: /home/arnaud/src/mc_rtc/build/tests/test_mc_rbdyn
Test timeout computed to be: 1500
Running 2 test cases...
../tests/test_mc_rbdyn.cpp(93): error: in "TestRobotZMPSimple": Error in Robot::zmp computation with leftFootRatio=
Expected: 0.0443193  0.109674  0
Got:      0.0476033  0.107422  0
../tests/test_mc_rbdyn.cpp(109): error: in "TestRobotZMPSimple": Error in Robot::zmp computation with leftFootRatio=
Expected: 0.0443574 -0.102838   0
Got:      0.0451703 -0.0954261  0
../tests/test_mc_rbdyn.cpp(115): error: in "TestRobotZMPSimple": exception std::runtime_error expected but not raised
```
The calibration parameter

- The first two are most likely due to the offsets in sensor position and
force computed by the calibration.
- The last error is due to the calibrated mass of 1.48995kg, that is
higher than the threshold for minimal force on the ZMP (1N), and thus
the minimal force threshold test will pass even though the function was
called with zero force (the Robot::netWrench computation uses the
wrenchWwrenchWithoutGravity function).

-----------------
TODO (important):
-----------------
We should fully investigate the corectness of using the calibrated
wrench for computing the ZMP. I have a feeling that this is incorrect,
intuitively when the feet are in contact, it doesn't make sense to cancel the
link's weight (the pose offset for the sensor does, and maybe the force
offset)