[Logger] Add a warning if the logging queue is outrun
Also fix a memory leak in that case. This was noticed by running a
"greedy" ticker in valgrind but I don't think it happens in practice as
the queue is relatively large and the consumer thread runs frequently.

It can also happen when running tests on a real-time kernel as the
ticker can be very fast in those cases and the consumer thread runs
infrequently in comparison (every 500us at best). Reported by @chappk