[GUI] Allow new plots to be added after the initial addPlot call
The main purpose of this patch is to allow to build a plot dynamically.

Plots cannot be removed dynamically as we lose much of the plot
information after it has been built. We could keep track of enough
information to do it but that would be more costly in the general case
with little benefits.