[GUI] Allow dynamic elements in form
This is particularly targeted at allowing changing online values that
need to be changed simultaneously (e.g. stiffness/damping) but might
also change through code.

This is done transparently for the existing clients but might require a
new "locking" mechanism in the client to allow editing forms where the
provided values change often.

This breaks some existing syntax for adding form elements as those
elements are added through a function call instead of an object. Forms are not
used much at the moment and the syntax is more in-line with other GUI
code so this should not break a lot of existing code.