[mc_control] Implement multi-robot aware init and reset
Multi-robot init is an extension of the existing init method. The
default "empty" case is also a reasonable init call for starting an
open-loop global controller instance.

reset allows to reset the controller without re-creating the global
controller instance. This allows to implement a more efficient way to
restart a controller, especially in multi-threaded scenarios.