[mc_rtc] YAML improvement
- Handle simple merge keys
- Add logic around bool conversion to ignore y/n

Merge keys exist in YAML 1.1 but were removed in YAML 1.2. They are not
supported by yaml-cpp which we use internally for all YAML handlings
(note that internally all YAML inputs are converted to JSON). However,
they can be very useful for us (as demonstrated in the unit test):

```yaml
default-col: &default-col
  iDist: 0.05
  sDist: 0.01
  damping: 0
collisions:
- b1: body1
  b2: body2
  <<: *default-col
... Repeat with a lot of collisions
```

However, YAML 1.1 also allows the following:

```yaml
first: &first
  a: 0
  b: 1
second: &second
  a: 1
  c: 0
third:
  <<: [*first, *second]
```

This is not supported here as it would be more complicated and the
semantic is not very clear.