[mc_rtc] Introduce mc_rtc::shared
This is a simple around std::enable_shared_from_this that allows
implicit conversion from a reference to an object to a shared pointer to
this object.

Class that derives from this type should only be instantiable as
shared_ptr to avoid misuse.

This allows manipulating references in the API whose lifetime can be
extended on demand.