Make mc_solver::QPSolver an abstract class
The class exposes common functionalities for different backends to add
and remove constraints and tasks implemented using mc_rtc abstractions
(mc_tasks::MetaTask and mc_solver::ConstraintSet).

A specialization is implemented for Tasks in mc_solver::TasksQPSolver,
it exposes all functionalities from the previous incarnation of
mc_solver::QPSolver

However, mc_solver::msg::QPResult and other messages have been removed.
Interfaces that are using it should be adapted (mc_openrtm is the only
one that I know of)

Tasks and constraints are implemented to handle different backends using
type erasure.

General users don't have to implement those tricks if they are writing
their tasks or constraint for a specific backend, they only need to make
sure the backend that is being used match their expectations.

mc_tasks::AddRemoveContactTask is an example of adapting a task to the
changes with minimal changes.

In order not to change the existing signatures the QPSolver class
carries a static thread_local variable that is used by ConstraintSet and
MetaTask to identify the backend at creation (where actual types are
erased) and confront it with the backend at insertion or removal.

Added
--

- [mc_tasks] ImpedanceTask can use a Frame in the constructor (forgotten
  in #247)

API breaks
--

mc_solver
==

- Adding Tasks-based constraints and tasks should now be done after
  casting the mc_solver::QPSolver to an mc_solver::TasksQPSolver, one
  should make sure the solver backend is Tasks before doing so
- More generally, accessing Tasks specific API from mc_solver::QPSolver
  requires such a cast
- addToSolver(mbs, tasks::qp::QPSolver) and
  removeFromSolver(tasks::qp::Solver) no longer exist in ConstraintSet.
  Instead, one should implement addToSolverImpl(mc_solver::QPSolver) and
  removeFromSolverImpl(mc_solver::QPSolver)
- Constraints are no longer default-constructible
- ContactContstraint does not accept a dynamics parameter anymore, the
  positive lambda constraint is systematically added instead
- contactConstr/posLambdaConstr are no longer accessible from
  ContactConstraint
- motionConstr is no longer accessible from DynamicsConstraint
- RobotEnvCollisionConstraint removed (unused)

mc_tasks
==

- TrajectoryTaskGeneric is no longer templated
- TrajectoryTaskGeneric::finalize is templated by the error type
- TrajectoryTaskGeneric error, trajectory and selector are type-erased
- Removed mc_tasks::extraStiffness

mc_control
==

- MCController now accepts a QPSolver::Backend parameter
- Built-in constraints are now unique_ptr and not values, unless a
  controller is overwriting them, this does not break existing code
  thanks to a new overload for `addConstraintSet` in QPSolver
- [FSM] SlidingFootContact state has been removed