[mc_control] Avoid segfaults due to invalidated configuration objects
With the default robots' specific configuration loading mechanism,
robot's specific configuration are loaded into the robots map in the
configuration. This can create new key in this map (especially if the module name
does not match the name of the robot). This can the Configuration that
we are iterating upon.

To overcome this issue we capture the list of keys first and get the
configuration object one by one.

Closes #342