[mc_rbdyn] Implement `addForceSensor`
This function allows to insert a new force sensor into a robot at
run-time.

This also updates the frames of every robot that uses the same data.
This means that adding a force sensor to the control robot will not only
update the frames of the control robot but also the real/observed robot
and their outputs counterpart.

The sensor calibration file is not searched, the usual scenario where
this might be used implies that calibration files do not exist or do not
reside in the usual location so searching for them would be wasteful.