[bugfix] Ensure that contacts have a name before inserting into contact set
This ensures uniqueness of contacts.
Otherwise adding a contact without r1/r2  and then re-adding the same
contact with r1/r2 set to the default robot leads to duplicated contacts

[bugfix] Ensure that contacts have a name before inserting/removing into contact set

This ensures uniqueness of contacts.
Otherwise adding a contact without r1/r2  and then re-adding the same
contact with r1/r2 set to the default robot leads to duplicated contacts