[mc_rbdyn] Allow nested calls to RobotLoader::get_robot_module
The introduction of RobotModule::connect made it interesting to compose
robot modules. In order to expose new modules that are composed of
multiple other existing modules, it is necessary to allow loading other
robot modules while creating a new one.

This effectively allows calling get_robot_module from within the
create() function of a new robot module library. To do so we unlock the
mutex just before the actual call to create().

This may have unintended side effects.